updated post gen hook to account for the fact that options under produces/exe/* don't have a test/ directory to remove

+++ hooks/post_gen_project.py
@@ -40,7 +40,8 @@ def remove_external():
 
 def remove_test():
     p = Path("test")
-    shutil.rmtree(p)
+    if p.is_dir():
+        shutil.rmtree(p)
 
 def main():
